change tester from main to test package
