+++ tester/it_works/expected_output
@@ -1,3 +1,1 @@
 It really does work!
-aa alskdfj
-laskd
