--- tester/it_works/output
@@ -1,1 +1,0 @@
-It really does work!
